backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
chromadb
pymupdf
pdfplumber
python-multipart
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
pytest
